Attack - Joystick has been added - Code is cleaner and simpler
The player uses weapons with a keyboard and/or a joystick. You can switch with space/square
The spray weapon is only usable for a certain number of seconds.
Cleaned code and made variables more understandable.
Added inputs to the manager for the joystick.

MAJOR CHANGES
WeaponManager was unnecessarily long cause I didn't know how to go back on the array once you finish it.
PlayerAttackDistance controls how the mouse and/or the joystick attack. At the moment, the mouse uses the new input while the joystick uses the old one.